requirements.txt
pybis==1.37.0
ollama-python==0.4.7
httpx==0.24.0